1. Move _create_azconfig_pipeline out of _generated 2. Cleanup imports in azure_configuration_requests
