+++ azure-configurationservice/azure/configuration/_generated/azure_configuration_client_imp.py
@@ -19,16 +19,6 @@ import uuid
 from . import models
 from .. import utils
 
-from .. import azure_configuration_requests
-from msrest.pipeline.requests import (
-    PipelineRequestsHTTPSender,
-    RequestsPatchSession
-)
-from msrest.pipeline import Request, Pipeline
-from msrest.universal_http.requests import (
-    RequestsHTTPSender,
-)
-
 class AzureConfigurationClientImpConfiguration(AzureConfiguration):
     """Configuration for AzureConfigurationClientImp
     Note that all parameters used to create this instance are saved as instance
@@ -68,26 +58,11 @@ class AzureConfigurationClientImp(object):
 
         self.config = AzureConfigurationClientImpConfiguration(connection_string)
         self._client = ServiceClient(credentials, self.config)
-        self._client.config.pipeline = self._create_azconfig_pipeline()
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
         self.api_version = '1.0'
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
-
-    def _create_azconfig_pipeline(self):
-        policies = [
-            self.config.user_agent_policy,  # UserAgent policy
-            RequestsPatchSession(),         # Support deprecated operation config at the session level
-            self.config.http_logger_policy  # HTTP request/response log
-        ]
-
-        policies.insert(1, azure_configuration_requests.AzConfigRequestsCredentialsPolicy(self.config))  # Set credentials for requests based session
-
-        return Pipeline(
-            policies,
-            PipelineRequestsHTTPSender(RequestsHTTPSender(self.config))  # Send HTTP request using requests
-        )
 
     def list_key_values(
             self, label=None, key=None, accept_date_time=None, fields=None, custom_headers=None, raw=False, **operation_config):
